requirements.txt
Django==2.2.8
django-debug-toolbar==2.1
django-split-settings==1.0.0
pkg-resources==0.0.0
python-decouple==3.3
python-memcached==1.59
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
